merge: move cli
+ MANIFEST.in
