+++ MANIFEST.in
@@ -1,3 +1,1 @@
-global-exclude *.pyc
-global-exclude __pycache__
 recursive-include brainsets_pipelines *
