--- brainsets/MANIFEST.in
@@ -1,1 +1,0 @@
-recursive-include brainsets_pipelines *
