+++ torch_brain/pipeline/_cli/cli_config.py
@@ -5,7 +5,7 @@ from prompt_toolkit import prompt
 from prompt_toolkit.shortcuts import CompleteStyle
 from prompt_toolkit.completion import PathCompleter
 
-from brainsets.config import CONFIG_FILE, load_config, save_config
+from torch_brain.pipeline.config import CONFIG_FILE, load_config, save_config
 
 from .utils import expand_path
